requirements.txt
aiosqlite==0.19.0
alembic==1.12.1
amqp==5.2.0
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.29.0
billiard==4.2.0
celery==5.3.5
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
et-xmlfile==1.1.0
fastapi==0.104.1
greenlet==3.0.1
h11==0.14.0
idna==3.4
kombu==5.3.4
Mako==1.3.0
MarkupSafe==2.1.3
mypy==1.6.1
mypy-extensions==1.0.0
openpyxl==3.1.2
pdf2image==1.16.3
Pillow==10.1.0
prompt-toolkit==3.0.41
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic_core==2.10.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
python-slugify==8.0.1
redis==5.0.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
text-unidecode==1.3
typing_extensions==4.8.0
tzdata==2023.3
uvicorn==0.24.0.post1
vine==5.1.0
wcwidth==0.2.11
